+++ src/index.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="utf-8">
   <title>ContentPlayer</title>
-  <base href="/">
+  <base href="/"> 
 
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
